ProgressReporters now works with ints instead of floats

+++ LibDayDataExtractor/LibDayDataExtractor/Progress/IProgressReporter.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace LibDayDataExtractor.Progress
 {
-    public interface IProgressReporter : IProgress<float>
+    public interface IProgressReporter : IProgress<int>
     {
-        float Progress { get; }
+        int Progress { get; }
     }
 }
